Add ability to use a local aria-repo
To save the time/bandwidth of cloning everytime

+++ scripts/update-reference.js
@@ -6,23 +6,27 @@ const mustache = require('mustache');
 const np = require('node-html-parser');
 
 const tmpPath = path.join(__dirname, "tmp");
+let repoBasePath = tmpPath;
 const cloneURL = 'https://github.com/w3c/aria-practices'
 
 const args = require('minimist')(process.argv.slice(2), {
   alias: {
     h: 'help',
+    r: 'aria-practices-repo',
   },
 });
 
 if (args.help) {
   console.log(`
 Default use:
-  node update-reference.js exampleName
-    Will copy the latest example html from aria-practices repo into this repo.
+  node update-reference.js exampleName [-a ariaPracticesDirectory]
+    It will copy the latest example html from the aria-practices repo into this repo.
+    It will clone the latest aria-practice directory or use an existing clone.
     The example html url for the example needs to be specificed in the references.csv file for the example named.
     The copied reference will be placed in a dated folder to provide a form of version control for the refernece files.
     The new reference will only be used with the reference.csv file is updated to point to the new reference folder.
   Arguments:
+    -r, --aria-practices-repo
     -h, --help
        Show this message.
 `);
@@ -83,27 +87,32 @@ async function copyExampleToRepo(exampleName) {
       console.log('`example` must be defined in references.csv with the format `https://w3c.github.io/aria-practices/examples/<PATH_TO_EXAMPLE>.html`');
       process.exit();
     }
-    await fse.remove(tmpPath);
-    console.log('Cloning the aria-practice repo.');
-    await git.Clone(cloneURL, tmpPath);
-    console.log(`Locating the matching example file ${examplePath}.`);
 
+    if (!args.r) {
+      await fse.remove(tmpPath);
+      console.log('Cloning the aria-practice repo.');
+      await git.Clone(cloneURL, tmpPath);
+    } else {
+      repoBasePath = args.r
+    }
+
+    const htmlFileAbsolute = path.join(repoBasePath, examplePath);
+    console.log(`Locating the matching example file ${htmlFileAbsolute}.`);
     try {
-      fse.statSync(path.join(tmpPath, examplePath));
+      fse.statSync(htmlFileAbsolute);
     }
     catch (err) {
-      console.log("The example html '" + examplePath + "' does not exist. Please enure the current example html utl is in the references.csv file.");
+      console.log("The example html '" + htmlFileAbsolute + "' does not exist. Please enure the current example html utl is in the references.csv file.");
       process.exit();
     }
 
-    const htmlFileAbsolute = path.join(tmpPath, examplePath);
     const currentDateTime = new Date();
     const formattedDateTime = currentDateTime.getFullYear() + "-" + (currentDateTime.getMonth() + 1) + "-" + currentDateTime.getDate() + "_" + currentDateTime.getHours() + + currentDateTime.getMinutes() + currentDateTime.getSeconds();
     const referenceDir = path.join(tmpPath, '..', '..', 'tests', exampleName, 'reference', formattedDateTime);
     await fse.ensureDir(referenceDir);
     const filterFunc = (src) => { return (src.indexOf('.html') == -1 || src == htmlFileAbsolute) };
     console.log('Coping assets to timestamped local directory.\n\n');
-    await fse.copy(path.join(tmpPath, 'examples', htmlFileAbsolute.split('examples' + path.sep)[1].split(path.sep)[0]), referenceDir, { filter: filterFunc});
+    await fse.copy(path.join(repoBasePath, 'examples', htmlFileAbsolute.split('examples' + path.sep)[1].split(path.sep)[0]), referenceDir, { filter: filterFunc});
     const referenceHtml = locateFile(referenceDir, path.basename(examplePath));
     const referenceHtmlPath = path.join('reference', referenceHtml.split('reference' + path.sep)[1]);
     console.log(`Reference file created at tests/${exampleName}/${referenceHtmlPath}.\nTo switch the test to run the updated reference:\n\t1. Commit this change\n\t2. Update ${path.join('tests', exampleName, 'data', 'reference.csv')} with the reference ${referenceHtmlPath}\n\t3. Open the html file and edit it to only include the example. The title, imported assets, h1 with the example name, and the div with the actual example (Usually #ex1) need to be preserved, but everything else can be removed.`);
